senergy-ml/requirements.txt
# Core ML dependencies
tensorflow==2.15.0
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2

# Flask API
flask==3.0.0
flask-cors==4.0.0

# Firebase
firebase-admin==6.3.0

# Utilities
python-dotenv==1.0.0
pandas==2.1.4

# for visualization and monitoring
matplotlib==3.8.2
seaborn==0.13.0